Task: Finalize C view model and controllers. Closes #13

+++ products/models/albertheijnProducts.py
@@ -1,10 +1,9 @@
 from django.db import models
-from django.utils import timezone
-from products.models import BaseProduct
+from .baseProducts import BaseProduct
 
 
 class AlbertHeijnProduct(BaseProduct):
-    product_id = models.PositiveIntegerField(verbose_name="product id", unique_for_date=True)
+    product_id = models.BigIntegerField(verbose_name="product id", unique_for_date=True)
 
     def __str__(self):
         value = 'product_id: {product_id}'.format(product_id=self.product_id) + ', entry_date:' + str(self.entry_date)
